backend/NN/requirements.txt
fastapi
uvicorn[standard]
jinja2
pillow
numpy
torch
torchvision
timm
python-multipart